+++ src/App.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { Routes } from './pages/Routes'
+
+function App() {
+    return (
+        <Routes />
+    );
+}
+
+export default App;
